+++ src/fs/create.js
@@ -1,9 +1,9 @@
-import { __dirname, errMessage } from './utils.js';
+import { getPath, errMessage } from '../utils.js';
 import { join } from 'node:path';
 import { writeFile } from 'node:fs/promises';
 
 const create = async () => {
-  const pathFile = join(__dirname, 'files', 'fresh.txt');
+  const pathFile = join(getPath(import.meta.url), 'files', 'fresh.txt');
   const fileContent = 'I am fresh and young';
     try {
       await writeFile(pathFile, fileContent, { flag: 'wx' });
