+++ src/fs/list.js
@@ -1,10 +1,10 @@
-import { __dirname, errMessage } from './utils.js';
+import { getPath, errMessage } from '../utils.js';
 import { join } from 'node:path';
 import { readdir } from 'node:fs/promises';
 
 const list = async () => {
     try {
-      const pathFile = join(__dirname, 'files');
+      const pathFile = join(getPath(import.meta.url), 'files');
       const listFiles = await readdir(pathFile);
       console.log(listFiles);
     } catch (error) {
