+++ src/fs/rename.js
@@ -1,10 +1,10 @@
-import { __dirname, errMessage, exist } from './utils.js';
+import { getPath, errMessage, exist } from '../utils.js';
 import { join } from 'node:path';
 import { rename as renameFile } from 'node:fs/promises';
 
 const rename = async () => {
-  const pathOldFile = join(__dirname, 'files', 'wrongFilename.txt');
-  const pathNewFile = join(__dirname, 'files', 'properFilename.md');
+  const pathOldFile = join(getPath(import.meta.url), 'files', 'wrongFilename.txt');
+  const pathNewFile = join(getPath(import.meta.url), 'files', 'properFilename.md');
   
   try {
     if (await exist(pathNewFile)) throw new Error(errMessage);
